# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bowtie API
  version: 0.1.0
  description: Live TV streaming API for HDHomeRun tuners.
paths:
...
  /api/v1/me/recents:
    get:
      summary: Channels the current user watched recently
      description: |
        Newest first; enabled channels only. A channel is recorded after 30 s
        of watching, so zapping through doesn't add it.
      operationId: listRecents
      tags: [me]
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 20
            default: 8
      responses:
        '200':
          description: Recently watched channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecentChannel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
    delete:
      summary: Clear the current user's watch history
      operationId: clearRecents
      tags: [me]
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Cleared
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/iptv/{key}/playlist.m3u:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      summary: M3U playlist of the feed account's channels
      operationId: iptvPlaylist
      tags: [iptv]
      responses:
        '200':
          description: Extended M3U (tvg-id matches guide.xml channel ids)
          content:
            audio/x-mpegurl:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/iptv/{key}/guide.xml:
    parameters:
      - name: key
        in: path
        required: true
        schema:
          type: string
    get:
      summary: XMLTV guide (next 72 hours) for the feed account's channels
      operationId: iptvGuide
      tags: [iptv]
      responses:
        '200':
          description: XMLTV document
          content:
            application/xml:
              schema:
                type: string
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/v1/channels:
    get:
      summary: List enabled channels for the viewer
      operationId: listChannels
      tags: [channels]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Enabled channels with optional EPG logo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ViewerChannel'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    RecentChannel:
      type: object
      required: [channelId, guideNumber, name, logoUrl, watchedAt]
      properties:
        channelId:
          type: integer
          format: int64
        guideNumber:
          type: string
        name:
          type: string
        logoUrl:
          type: string
        watchedAt:
          type: string
          format: date-time
    ViewerChannel:
      type: object
      required: [id, guideNumber, name, logoUrl, reception, favorite]
      properties:
        favorite:
          type: boolean
          description: The caller has starred this channel
        id:
          type: integer
          format: int64
        guideNumber:
          type: string
        name:
          type: string
        logoUrl:
          type: string
          description: Icon URL from mapped EPG channel, or empty
        reception:
          type: string
          enum: [ok, noSignal, unknown]
          description: |
            The channel's last tune outcome, learned from real tunes (the
            HDHomeRun reports signal only for a channel it is tuned to).
            "noSignal" means the tuner last answered 806/807; "unknown" means
            not tuned since the server started.
        receptionCheckedAt:
          type: string
          format: date-time
          description: When reception was last learned; absent when unknown.
...
  responses:
    BadRequest:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication failed or missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'